refactor(easing): split 1010-line file into 5 focused modules
- types.rs: Core types and enums
- functions.rs: Easing function implementations
- helpers.rs: Constructor functions
- tests.rs: All unit tests with #![allow(unused_imports)]
- mod.rs: Re-exports for backward compatibility

Closes #13
